requirements.txt
python-binance==1.0.17
customtkinter==5.2.1
python-dotenv==1.0.1
